Collisions for Beaming Up
It seems like in order for the beam up to work it has to be a physical collision. This is fine for now, but it might need to be changed later. Also, you may need to make another script for actually receiving the object, as touching it otherwise will still count, regardless of whether or not the tractor beam is active.